+++ src/drone.c
@@ -75,21 +75,24 @@ int main(int argc, char *argv[])
     log_config(FILENAME_LOG, LOG_DEBUG);
 
     WritePid();
-    kill(watchdogPid, SIG_WRITTEN);
+    if (watchdogPid != -1) kill(watchdogPid, SIG_WRITTEN);
     
     //Protocol messages
     Msg_int msg_int_in;
     Msg_float msg_float_in, msg_float_out;
 
-    //Signal from watchdog
-    struct sigaction sa_ping;
-    sa_ping.sa_handler = ping_handler;
-    sa_ping.sa_flags = SA_RESTART;
-    sigemptyset(&sa_ping.sa_mask);
-    
-    if (sigaction(SIG_PING, &sa_ping, NULL) == -1) {
-        perror("Error in ping_handler");
-        exit(EXIT_FAILURE);
+    if (watchdogPid != -1)
+    {
+        //Signal from watchdog
+        struct sigaction sa_ping;
+        sa_ping.sa_handler = ping_handler;
+        sa_ping.sa_flags = SA_RESTART;
+        sigemptyset(&sa_ping.sa_mask);
+        
+        if (sigaction(SIG_PING, &sa_ping, NULL) == -1) {
+            perror("Error in ping_handler");
+            exit(EXIT_FAILURE);
+        }
     }
